Resources: correct string encoding conversion
The UTF8 to UTF16 conversion was error-prone as it truncated the top-bit
and used the wrong value for the character count for the string
creation.  This uses Foundation to perform the string conversion and
counts the new character length.

The string conversion of the character to UTF16 also houses another
failure point by not ensuring that we do not require a surrogate pair
for the character encoding.  However, leave that issue in place but
remove the use of the `withWideChars` conversion.  In the process we
avoid an allocation of a `String` by directly using the UTF16 view and
stripping the first integral value (potentially half of a surrogate
pair).