change code generation to generate files per type (#224)
This change changes the structure of the generated files, so that there is one file-per-type. There is also a mode to generate files based on category (classes/enums/delegates/etc).

The reason for this change is there are lots of `internal` APIs which are referenced between the different ABI/Impl files and we will be able to make these `fileprivate`, resulting in fewer dll exports being used when using a build system like SPM